Retry container rescheduling
Introduce --reschedule-retry and --reschedule-retry-interval command
line options to specify the number of times the primary manager
will attempt the reschedule a container and the duration to wait
in between rescheduling attempts, respectively.

By default, the manager will not attempt to retry failed rescheduling
attempts.

A few scenarios where the addition of retries can help with cluster
robustness:
* The attempt to create or start the rescheduled container fails due
  to a transient error in the docker engine (e.g. docker/docker#20398).
* The container being rescheduled had a lease on a global Docker Volume
  (e.g. a network accessible Volume provided by a third-party
   VolumeDriver plugin) that prevents the rescheduled container from
   mounting the volume until the lease expires.

Signed-off-by: Daniel Ferstay <dan@cohodata.com>